refactor: centralize daemon bootstrap trace types
