+++ crates/atm-daemon-client/Cargo.toml
@@ -14,7 +14,6 @@ atm-core = { package = "agent-team-mail-core", path = "../atm-core", version = "
 atm-storage = { path = "../atm-storage", version = "1.3.2-beta-30" }
 fs2 = "0.4"
 interprocess = "2.4"
-serde = { workspace = true }
 serde_json.workspace = true
 tracing = "0.1"
